+++ omg-react-app/src/components/Button.jsx
@@ -1,14 +1,8 @@
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
 
 import styles from '../styles/Button.module.css'
 
 function Button(props) {
-    const navigate = useNavigate();
-
-    const goToMain = () => {
-        navigate('/main')
-    };
 
     if (props.type === "big") {
         return (
@@ -25,17 +19,6 @@ function Button(props) {
             <button className={styles["prev"]}>
                 <span className={styles["prev-span"]}>{props.content}</span>
             </button>
-        )
-    }
-
-    if (props.type === "login") {
-
-        return (
-            <div className={styles["login-div"]}>
-                <button onClick={goToMain} className={styles["login"]}>
-                    <span className={styles["login-span"]}>{props.content}</span>
-                </button>
-            </div>
         )
     }
